Move tests to correct packages
