+++ junit-jupiter-theories/src/test/java/org/junit/jupiter/theories/DataPointRetrieverTests.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junit.jupiter.theories.util;
+package org.junit.jupiter.theories;
 
 import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
